% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alps}
\alias{Alps}
\title{Alpine region}
\format{
An sf object :
\describe{	
  \item{name}{European Alps}
  \item{geometry}{sfc_MULTIPOLYGON geometry}
}
}
\source{
Alpine Convention
}
\usage{
Alps
}
\description{
A spatial dataset (class "sf", "data.frame") containing the borders of the European Alps region,
retrieved from the Alpine Convention website (https://www.alpconv.org/fr/page-daccueil/).
Used for examples and testing within the envar package.
}
\keyword{datasets}
