% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridity.R
\name{aridity}
\alias{aridity}
\title{Download and process Global Aridity Index and Potential Evapotranspiration (ET0)}
\usage{
aridity(x, vars, ...)
}
\description{
Download and process Global Aridity Index and Potential Evapotranspiration (ET0)
}
