% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biooracle.R
\name{biooracle}
\alias{biooracle}
\title{Download Bio-ORACLE marine data}
\usage{
biooracle(
  x,
  vars,
  realm = "surface",
  years = "2000-2010",
  ssp = NULL,
  algorithm = "mean",
  ...
)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations.}

\item{vars}{Character vector of one or more variables or synonyms to download.}

\item{realm}{Character. One of "surface" (default), "benthic_minimum", "benthic_average", or "benthic_maximum".}

\item{years}{Character. Decade range (e.g., "2000-2010" up to "2090-2100").}

\item{ssp}{numeric or character. Shared Socioeconomic Pathway (119, 126, 245, 370, 460, 585).}

\item{algorithm}{Character. Statistic to apply (max, mean, min, ltmax, ltmin, range). Default "mean".}

\item{...}{Additional arguments.}
}
\description{
This function downloads, processes, and extracts variables from the
Bio-ORACLE v3.0 dataset.
}
\details{
Available variables (working synonyms in parentheses and ""):

1 - "thetao" (Ocean temperature; [ºC]) ("temperature", "temp", "sea temperature")

2 - "so" (Salinity; [-]) ("sal", "salt", "saltiness")

3 - "sws" (Sea water velocity; [m.s-1]) ("velocity", "current speed", "speed")

4 - "swd" (Sea water direction; [degree]) ("direction", "current direction")

5 - "no3" (Nitrate; [mmol . m-3]) ("nitrate")

6 - "po4" (Phosphate; [mmol . m-3]) ("phosphate")

7 - "si" (Silicate; [mmol . m-3]) ("silicate", "silicon")

8 - "o2" (Dissolved molecular oxygen; [mmol . m-3]) ("oxygen", "dissolved oxygen", "o2")

9 - "dfe" (Iron; [mmol . m-3]) ("iron", "fe")

10 - "phyc" (Primary productivity; [mmol . m-3]) ("productivity", "pp", "primary production")

11 - "ph" (pH; [-]) ("acidity")

12 - "chl" (Chlorophyll; [mg . m-3]) ("chlorophyll", "chla")

13 - "sithick" (Sea ice thickness; [m]) ("ice thickness")

14 - "siconc" (Sea ice cover; [Fraction]) ("ice cover", "sea ice")

15 - "clt" (Cloud cover; [%]) ("cloud", "clouds")

16 - "mlotst" (Mixed layer depth; [m]) ("mld", "mixed layer")

17 - "tas" (Air temperature; [ºC]) ("air temperature", "air temp")

18 - "par" (Photosynt. Avail. Radiation; [E.m-2.day-1]) ("light", "radiation")

19 - "kdpar" (Diffuse attenuation; [m-1]) ("attenuation", "turbidity")

20 - "bathymetry" (Bathymetry; [m]) ("depth", "elevation", "altitude")

21 - "slope" (Topographic slope; [-]) ("topographic slope")

22 - "aspect" (Topographic aspect; [-]) ("topographic aspect")

23 - "tpi" (Topographic position index; [-]) ("topographic position index")

24 - "tri" (Terrain ruggedness index; [-]) ("terrain ruggedness index", "ruggedness")

Citation:
Assis, J., et al. (2024). "Bio-ORACLE v3.0. Pushing marine data layers to the CMIP6 Earth system models of climate change research." 
Global Ecology and Biogeography. DOI: {.url https://doi.org/10.1111/geb.13813}
}
