% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envar-package.R
\docType{package}
\name{envar-package}
\alias{envar}
\alias{envar-package}
\title{envar: Download and Process Environmental Variables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a unified interface to download, harmonise and extract a wide range of environmental and socio-economic variables from established open data sources (such as 'WorldClim', 'CHELSA', 'Bio-ORACLE', 'SoilGrids' and others) for use in macroecology and biogeography. The package handles spatial subsetting to a study area, reprojection to a common coordinate reference system, and extraction of values at sampling points, so that predictors from heterogeneous sources can be assembled within a single reproducible workflow. Helper functions for collinearity checking and variable exploration are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://animalbiodiversitylab.github.io/envar/}
  \item \url{https://github.com/animalbiodiversitylab/envar}
  \item Report bugs at \url{https://github.com/animalbiodiversitylab/envar/issues}
}

}
\author{
\strong{Maintainer}: Andrea Simoncini \email{simonciniandre@gmail.com} (\href{https://orcid.org/0000-0002-9856-6111}{ORCID})

Authors:
\itemize{
  \item Michele Bertoncini (\href{https://orcid.org/0000-0002-1866-0045}{ORCID})
  \item Andrea Cerofolini (\href{https://orcid.org/0009-0009-4439-3014}{ORCID})
  \item Andrea Dalpasso (\href{https://orcid.org/0000-0003-1471-789X}{ORCID})
  \item Mattia Falaschi (\href{https://orcid.org/0000-0002-4511-4816}{ORCID})
  \item Gentile Francesco Ficetola (\href{https://orcid.org/0000-0003-3414-5155}{ORCID})
  \item Elia Lo Parrino (\href{https://orcid.org/0000-0001-9619-9568}{ORCID})
}

}
\keyword{internal}
