% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esalandcover.R
\name{esalandcover}
\alias{esalandcover}
\title{Download and process Global 1 km Land Cover variables}
\usage{
esalandcover(x, vars, ...)
}
\arguments{
\item{x}{A `SpatRaster`, `SpatVector`, or `sf` object defining the area or
locations for extraction.}

\item{vars}{Character vector of variables, supplied as canonical codes or
friendly names.}

\item{...}{Reserved for future use.}
}
\value{
- If `x` is a raster: a `SpatRaster` stack of processed layers.  
- If `x` contains points: a `data.frame` of extracted values.
}
\description{
`esalandcover()` downloads, processes, and extracts variables from the
**Global 1 km Land Cover** dataset. Each variable corresponds to a global 
raster representing a specific land cover class or diversity index derived 
from very high-resolution imagery.
}
\details{
The function allows users to input either:
- **canonical variable codes**, e.g. `"wetland"`, `"tree"`, `"shannon"`
- **human-readable names**, e.g. `"forest"`, `"agriculture"`, 
  `"surface water"`, `"simpson index"`, etc.

It automatically:
- downloads the selected files from Figshare,
- crops/resamples/masks to match a user-provided `SpatRaster`, **or**
- extracts values when `x` is point data.


## Citation
If you use this data, please cite:

**Lo Parrino E, Simoncini A, Ficetola GF, Falaschi M (2025)** #' *Global 1 km land cover for macroecological modelling from very high resolution imagery.* #' Figshare.  
https://figshare.com/s/4e7dee46628b530aee03


## Available variables

| Human-readable name          | Canonical variable code              |
|------------------------------|--------------------------------------|
| wetland / marsh              | wetland                              |
| bare ground / soil           | bare                                 |
| built / urban area           | built                                |
| cropland / agriculture       | cropland                             |
| grassland / meadow           | grass                                |
| ice / snow / glacier         | ice                                  |
| land percentage              | land_perc                            |
| mangrove                     | mangrove                             |
| moss / lichen                | moss                                 |
| shrub / scrub                | shrub                                |
| tree / forest                | tree                                 |
| water / surface water        | water                                |
| simpson index                | simpson                              |
| shannon index                | shannon                              |
| evenness index               | evenness                             |
}
