% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Europe}
\alias{Europe}
\title{Europe}
\format{
An sf object :
\describe{	
  \item{name}{Europe}
  \item{geometry}{sfc_MULTIPOLYGON geometry}
}
}
\source{
GADM
}
\usage{
Europe
}
\description{
Europe shapefile from the GADM, the Global Administrative Areas database.
}
\keyword{datasets}
