% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdppast.R
\name{gdppast}
\alias{gdppast}
\title{Download and process Historical Real GDP and Electricity Consumption}
\usage{
gdppast(x, vars, year, ...)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations for extraction, 
the reference system, and the buffer. 
Leave this empty and use `var_get()` to define parameters for download.}

\item{vars}{Character vector of variables to download ("gdp" or "electricity").}

\item{year}{Numeric vector of years to download. Available range: 1992-2019.}

\item{...}{Additional arguments (currently unused).}
}
\value{
If `var_get()` contained a raster/polygon/points with buffer: a `SpatRaster` stack of processed variables. If `var_get()` contained spatial points or data.frame of points without buffer: a `data.frame` of x, y, and extracted values.
}
\description{
This function downloads, processes, and extracts variables from the
global 1km gridded revised real gross domestic product and electricity 
consumption dataset (1992–2019).
}
\details{
\strong{Available variables} (working synonyms in parentheses):

\strong{Economic Metrics}
\itemize{
  \item "gdp" ("gross domestic product", "real gdp", "economy", "economic output", "gross product")
}

\strong{Energy Metrics}
\itemize{
  \item "electricity" ("electricity consumption", "energy", "energy consumption", "power", "ec", "electric")
}

\strong{Years available}
\itemize{
  \item 1992 to 2019.
}

\strong{Citation:}\cr
Chen J, Gao M, Cheng S et al (2022). "Global 1 km x 1 km gridded revised real gross domestic product and electricity consumption during 1992–2019 based on calibrated nighttime light data." Scientific Data 9, 202.
https://doi.org/10.1038/s41597-022-01322-5

Note: Please cite original sources of primary datasets where appropriate.
}
\examples{
\dontrun{
# Get GDP for 2000 and 2010
processed <- var_get(country= "Italy", crs=3035) \%>\% 
  gdppast(vars="gdp", year=c(2000, 2010))

# Get Electricity and GDP for 2019
processed <- var_get(country= "Vietnam") \%>\% 
  gdppast(vars=c("electricity", "gdp"), year=2019)
}
}
