% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marspec.R
\name{marspec}
\alias{marspec}
\title{Download and process MARSPEC ocean climate variables}
\usage{
marspec(x, vars, discover = TRUE, ...)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations for extraction.}

\item{vars}{Character vector of one or more variables to download and process.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A `SpatRaster` stack or a `data.frame` of extracted values.
}
\description{
This function downloads, processes, and extracts variables from the
MARSPEC (Marine Spatial Ecology) dataset. Each variable corresponds to a 
global raster representing geophysical or bioclimatic marine conditions.
}
\details{
Available variables (working synonyms in parentheses):

CORE GEOPHYSICAL:
1 - "bathymetry" ("depth", "seafloor", "altitude", "elevation")
2 - "biogeo01" ("east_west_aspect", "ew_aspect", "sin_aspect")
3 - "biogeo02" ("north_south_aspect", "ns_aspect", "cos_aspect")
4 - "biogeo03" ("plan_curvature", "plan_curv")
5 - "biogeo04" ("profile_curvature", "prof_curv")
6 - "biogeo05" ("distance_to_shore", "dist_shore", "coast_dist")
7 - "biogeo06" ("slope", "bathymetric_slope")
8 - "biogeo07" ("concavity", "convexity")

CORE BIOCLIMATIC:
9  - "biogeo08" ("mean_annual_sss", "sss_mean")
10 - "biogeo09" ("min_sss", "sss_min")
11 - "biogeo10" ("max_sss", "sss_max")
12 - "biogeo11" ("range_sss", "sss_range")
13 - "biogeo12" ("variance_sss", "sss_var")
14 - "biogeo13" ("mean_annual_sst", "sst_mean")
15 - "biogeo14" ("coldest_month_sst", "sst_coldest")
16 - "biogeo15" ("warmest_month_sst", "sst_warmest")
17 - "biogeo16" ("range_sst", "sst_range")
18 - "biogeo17" ("variance_sst", "sst_var")

SEA ICE:
19 - "ice01" to "ice12" ("january_ice", "february_ice", etc.) - Monthly sea ice concentration

MONTHLY CLIMATOLOGIES:
"sst01" to "sst12" (e.g., "january_sst", "february_sst")
"sss01" to "sss12" (e.g., "january_sss", "march_sss")

Citation:
Sbrocco, E. J., & Barber, P. H. (2013). MARSPEC: ocean climate layers for 
marine spatial ecology: Ecological Archives E094-086. Ecology, 94(4), 979-979.
DOI: https://doi.org/10.1890/12-1358.1
}
