% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
Re-exported from \pkg{dplyr}. See \code{dplyr::\link[dplyr:reexports]{\%>\%}}
for details. This lets users chain \code{envar} functions, e.g.
\code{par_set(country = "Italy") \%>\% worldclim(vars = "bio1")}.
}
\keyword{internal}
