% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Download and process Global Population Projections (SSP)}
\usage{
population(x, vars, year = 2020, ssp = 1, ...)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations for extraction, 
the reference system, and the buffer. 
Leave this empty and use `var_get()` to define parameters for download.}

\item{vars}{Character vector. Use "population" (or synonyms).}

\item{year}{Numeric vector. Years to download. Available from 2020 to 2100 in 5-year intervals (e.g., c(2020, 2050)).}

\item{ssp}{Numeric vector. Shared Socioeconomic Pathways to download (1, 2, 3, 4, or 5).}

\item{...}{Additional arguments (currently unused).}
}
\value{
If `var_get()` contained a raster/polygon/points with buffer: a `SpatRaster` stack of processed variables. If `var_get()` contained spatial points or data.frame of points without buffer: a `data.frame` of x, y, and extracted values.
}
\description{
This function downloads, processes, and extracts variables from the
Global Population Projections dataset (Wang et al., 2022).
It provides 1-km grid population distributions from 2020 to 2100 
under five Shared Socioeconomic Pathways (SSPs).
}
\details{
Available variables (working synonyms in parentheses):

Population Counts:

1 - "population" ("pop", "inhabitants", "residents", "people", "count", "census")

Use the arguments `year` and `ssp` to filter the specific data required.

Citation:

Wang, X., Meng, X. & Long, Y. (2022). "Projecting 1 km-grid population 
distributions from 2020 to 2100 globally under shared socioeconomic pathways." 
Sci Data 9, 563.
https://doi.org/10.1038/s41597-022-01675-x

Note: Please cite original sources of primary datasets where appropriate.
}
\examples{
\dontrun{
processed <- var_get(country = "Italy", crs = 3035) \%>\% 
  population(vars = "population", year = 2050, ssp = 2)
  }
}
