% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Apollo}
\alias{Apollo}
\title{Apollo butterfly occurrences}
\format{
A data.frame object :
\describe{	
  \item{X}{longitude in WGS84}
  \item{Y}{latitude in WGS84}
}
}
\source{
GBIF
}
\usage{
Apollo
}
\description{
A "data.frame" containing the occurrence data of Apollo butterfly (Parnassius apollo),
downloaded from the Global Biodiversity Information Facility website (www.gbif.org), on December 10, 2025.
Used for examples and testing within the envar package.
}
\keyword{datasets}
