% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamlandcover.R
\name{gcamlandcover}
\alias{gcamlandcover}
\title{Download and process Global Future Land Use/Cover layers (2020-2100)}
\usage{
gcamlandcover(x, vars = "landcover", ssp = 126, year = 2020, ...)
}
\arguments{
\item{x}{The output from `par_set()` defining the area or locations for extraction, 
the reference system, and the buffer.}

\item{vars}{Character. Currently unused/ignored as this function returns the 
landcover map defined by `year` and `ssp`, but kept for consistency. 
Default is "landcover".}

\item{ssp}{Numeric or Character. The SSP scenario code (126, 245, 370, 434, 585). 
Ignored if `year` is 2020.}

\item{year}{Numeric. The year of simulation (2020, 2030, 2050, 2070, 2100).}

\item{...}{Additional arguments (currently unused).}
}
\value{
If `par_set()` contained a raster/polygon/points with buffer: a `SpatRaster` stack.
If `par_set()` contained spatial points without buffer: a `data.frame`.
}
\description{
This function downloads, processes, and extracts simulated global land use 
and land cover (LULC) data for the period 2020-2100.
}
\details{
The data represents 1 km resolution LULC maps. The original data is in 
World Mercator projection and will be automatically reprojected to the 
CRS defined in `par_set()`.

\strong{Land cover codes}
\itemize{
  \item 1 - Cropland
  \item 2 - Forest
  \item 3 - Grassland
  \item 4 - Urban
  \item 5 - Barren
  \item 6 - Water
}

\strong{Available Years}
\itemize{
  \item 2020, 2030, 2050, 2070, 2100
}

\strong{Available SSPs (Shared Socioeconomic Pathways)}
\itemize{
  \item 126 (SSP1-2.6)
  \item 245 (SSP2-4.5)
  \item 370 (SSP3-7.0)
  \item 434 (SSP4-3.4)
  \item 585 (SSP5-8.5)
}

\strong{Citation:}\cr
Zhang T, Cheng C, Wu X (2023). "Mapping the spatial heterogeneity of global land use and land cover from 2020 to 2100 at a 1 km resolution." Scientific Data 10, 748.
https://doi.org/10.1038/s41597-023-02637-7
}
\examples{
\donttest{
# Get Baseline (2020)
processed <- par_set(country= "Italy", crs=4326) \%>\% 
  gcamlandcover(year = 2020)

# Get Future (SSP5-8.5 in 2050)
processed <- par_set(country= "Italy", crs=4326) \%>\% 
  gcamlandcover(ssp = 585, year = 2050)
}
}
