% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpfuture.R
\name{gdpfuture}
\alias{gdpfuture}
\title{Download and process Future GDP Projections (SSP1-5)}
\usage{
gdpfuture(x, vars, year, ssp, ...)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations for extraction, 
the reference system, and the buffer. 
Leave this empty and use `var_get()` to define parameters for download.}

\item{vars}{Character vector of variables to download (synonyms for "gdp").}

\item{year}{Numeric vector of years to download (decades, e.g., 2030, 2050).}

\item{ssp}{Numeric vector of SSP scenarios to download (values 1 to 5).}

\item{...}{Additional arguments (currently unused).}
}
\value{
If `var_get()` contained a raster/polygon/points with buffer: a `SpatRaster` stack of processed variables. If `var_get()` contained spatial points or data.frame of points without buffer: a `data.frame` of x, y, and extracted values.
}
\description{
This function downloads, processes, and extracts variables from the
global gridded GDP projections compatible with the five Shared Socioeconomic 
Pathways (SSPs) for the period 1850–2100.
}
\details{
Available variables (working synonyms in parentheses):

Economic Metrics:

1 - "gdp" ("gross domestic product", "future gdp", "economic projection", "economy", "ssp gdp")

Scenarios (SSP):
The `ssp` argument accepts integers 1 through 5, corresponding to:
SSP1: Sustainability
SSP2: Middle of the Road
SSP3: Regional Rivalry
SSP4: Inequality
SSP5: Fossil-fueled Development

Years available: Decadal intervals from 1850 to 2100 (e.g., 2020, 2030, 2040...).

Citation:

Murakami, D., Yoshida, T., & Yamagata, Y. (2021). "Gridded GDP projections 
compatible with the five SSPs (shared socioeconomic pathways)." 
Frontiers in Built Environment 7: 760306.
https://doi.org/10.3389/fbuil.2021.760306

Note: Please cite original sources of primary datasets where appropriate.
}
\examples{
\dontrun{
# Get GDP for SSP1 and SSP5 in 2050
processed <- var_get(country= "France", crs=3035) \%>\% 
gdpfuture(vars="gdp", year=2050, ssp=c(1, 5))

# Get time series for SSP2
processed <- var_get(country= "India") \%>\% 
gdpfuture(vars="gdp", year=c(2020, 2030, 2040), ssp=2)
  }
}
