% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil}
\alias{soil}
\title{Download and process Harmonized World Soil Database v2.0}
\usage{
soil(x, vars = NULL, ...)
}
\arguments{
\item{x}{The output from `var_get()` defining the area or locations for extraction, 
the reference system, and the buffer. 
Leave this empty and use `var_get()` to define parameters for download.}

\item{vars}{Character vector of one or more variables to download and process.
Defaults to "hwsd" if left empty.}

\item{...}{Additional arguments (currently unused).}
}
\value{
If `var_get()` contained a raster/polygon/points with buffer: a `SpatRaster` stack of processed variables.
If `var_get()` contained spatial points or data.frame of points without buffer: a `data.frame` of x, y, and extracted values.
}
\description{
This function downloads, processes, and extracts variables from the
Harmonized World Soil Database v2.0 (HWSD v2.0). The variable corresponds 
to a global raster file at 1 km resolution representing soil types.
}
\details{
\strong{Available variables} (working synonyms in parentheses):

\strong{Soil Data}
\itemize{
  \item 1 - "hwsd" ("soil", "type", "soiltype", "soil type")
}

\strong{Citation:}\cr
FAO, IIASA (2023). "Harmonized World Soil Database v2.0." Food and Agriculture Organization of the United Nations, Rome and International Institute for Applied Systems Analysis, Laxenburg, Austria.
https://www.fao.org/soils-portal/data-hub/soil-maps-and-databases/harmonized-world-soil-database-v20/en/
}
\examples{
\dontrun{
processed <- var_get(country= "Italy", crs=3035) \%>\% 
soil(vars=c("soil"))
  }
}
